+++ Form/Elements/Images.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace RevisionTen\CMS\Form\Elements;
+
+use RevisionTen\CMS\Form\Types\ImageType;
+use Symfony\Component\Form\Extension\Core\Type\CollectionType;
+use Symfony\Component\Form\FormBuilderInterface;
+
+class Images extends Element
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        parent::buildForm($builder, $options);
+
+        $builder->add('images', CollectionType::class, [
+            'label' => 'Images',
+            'entry_type' => ImageType::class,
+            'entry_options' => [
+                'attr' => [
+                    'class' => 'well',
+                ],
+            ],
+            'allow_add' => true,
+            'allow_delete' => true,
+        ]);
+    }
+}