+++ CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.8] - 2018-11-21
+## [1.3.9] - 2018-11-22
+### Added
+- Added image gallery element and basic ImageType
 ### Changed
+- Made CollectionType sortable
+
+## [1.3.8] - 2018-11-21
+### Added
 - Added configuration option for login code expiration
 
 ## [1.3.7] - 2018-11-21